fix comment component for search and submemmit

+++ frontend/src/features/comments/Comments.js
@@ -3,6 +3,7 @@ import axios from "axios";
 import { apiURL } from "../../util/apiURL";
 import "../../css/Comments.css";
 import { AuthContext } from "../../providers/AuthContext";
+import { useHistory } from "react-router-dom";
 
 export default function Comments({ post_id }) {
   const API = apiURL();
@@ -12,7 +13,8 @@ export default function Comments({ post_id }) {
   const [toggleDeleteComments, setToggleDeleteComments] = useState(false);
 
   const [content, setContent] = useState("");
-  const { currentUser } = useContext(AuthContext);
+  const { currentUser, token } = useContext(AuthContext);
+  const history = useHistory();
 
   const toggle = (state, setState) => {
     !state ? setState(true) : setState(false);
@@ -55,24 +57,36 @@ export default function Comments({ post_id }) {
 
   const displayComments = () => {
     let showComments = comments.map((comment, i) => {
-      if (currentUser.id === comment.author_id && toggleDeleteComments) {
-        return (
-          <div className="comment">
-            <p key={i} className="commentItem">
-              <strong>{comment.username}</strong> commented "{comment.content}"
-            </p>
-            {toggleDeleteComments ? (
-              <img
-                className="deleteBtn"
-                src="https://img.icons8.com/color/24/000000/delete-sign.png"
-                alt="delete"
-                onClick={() => {
-                  handleDelete(comment.id);
-                }}
-              />
-            ) : null}
-          </div>
-        );
+      if (currentUser) {
+        if (currentUser.id === comment.author_id && toggleDeleteComments) {
+          return (
+            <div className="comment">
+              <p key={i} className="commentItem">
+                <strong>{comment.username}</strong> commented "{comment.content}
+                "
+              </p>
+              {toggleDeleteComments ? (
+                <img
+                  className="deleteBtn"
+                  src="https://img.icons8.com/color/24/000000/delete-sign.png"
+                  alt="delete"
+                  onClick={() => {
+                    handleDelete(comment.id);
+                  }}
+                />
+              ) : null}
+            </div>
+          );
+        } else {
+          return (
+            <div className="comment">
+              <p key={i} className="commentItem">
+                <strong>{comment.username}</strong> commented "{comment.content}
+                "
+              </p>
+            </div>
+          );
+        }
       } else {
         return (
           <div className="comment">
@@ -92,7 +106,11 @@ export default function Comments({ post_id }) {
               src="https://img.icons8.com/ios/50/000000/send-comment.png"
               alt="add"
               onClick={() => {
-                toggle(toggleAddComment, setToggleAddComment);
+                if (currentUser && token) {
+                  toggle(toggleAddComment, setToggleAddComment);
+                } else {
+                  history.push("/login");
+                }
               }}
             />
             <img
@@ -100,7 +118,11 @@ export default function Comments({ post_id }) {
               src="https://img.icons8.com/ios-filled/50/000000/delete-chat.png"
               alt="delete"
               onClick={() => {
-                toggle(toggleDeleteComments, setToggleDeleteComments);
+                if (currentUser && token) {
+                  toggle(toggleDeleteComments, setToggleDeleteComments);
+                } else {
+                  history.push("/login");
+                }
               }}
             />
             {toggleAddComment ? (
@@ -150,7 +172,10 @@ export default function Comments({ post_id }) {
                   }}
                 />
                 <button type="submit" className="enterBtn">
-                  <img src="https://img.icons8.com/material-sharp/24/000000/enter-2.png" />
+                  <img
+                    src="https://img.icons8.com/material-sharp/24/000000/enter-2.png"
+                    alt="enter"
+                  />
                 </button>
               </form>
             ) : null}
